% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRESCAL.R
\name{PRESCAL_ALS}
\alias{PRESCAL_ALS}
\title{RESCAL with Projection using Alternate Least Squares}
\usage{
PRESCAL_ALS(X, rank, P, ...)
}
\arguments{
\item{X}{input list of length r , r being the number of relations.
Each element of the list must be a n by n matrix. n being the number of entities}

\item{rank}{rank of the factorization}

\item{P}{Projection matrix}

\item{maxIter}{Number of iterations, default 500}

\item{conv}{Tolerance parameter for the convergence, default 1e-4}

\item{lam.A}{Tuning parameter for A}

\item{lam.R}{Tuning parameter for R}

\item{A0}{Initial value for the n by rank matrix A}

\item{initA}{The initialization process of A : n by rank matrix. Can be either 'random' or 'EigenDecomp'. N
Doesn't need to specified if A0 is provided. Default is 'EigenDecomp'}

\item{compute_fit}{Specifies whether the convergence is based on taking the difference of
the norm of X[\link{k}] and AR[\link{k}]A^T. If TRUE it computes the difference,
if n is very large, it is recommended to set FALSE.}
}
\value{
The factorization A and a list of weights matrix R with length of the list as the length of X
}
\description{
Produce the the reduced rank factor \code{A} and the weight matrix \code{R} as a list for to each relation
}
\examples{
set.seed(12345)
n <- 4; p <- 2; d <- 2;
X1 <- sparseMatrix(i=c(1,3,4), j=c(3,2,4), x=1, dims=c(n,n))
X2 <- sparseMatrix(i=c(1,3,4,4), j=c(1,2,3,4), x=1, dims=c(n,n))
X  <- list(X1,X2)
result <- PRESCAL_ALS(X,p,P=matrix(rnorm(d*n),d,n), lam.A=1, lam.R=1, compute_fit=TRUE)
}
\author{
Salil Koner, Dhrubajyoti Ghosh \cr Maintainer: Salil Koner
\email{skoner@ncsu.edu}
}
